requirements.txt
python-dotenv
asyncpg
aiogram
mysql-connector-python